Corrige tipos dos IDs nas classes de repositorio (#3)

+++ src/main/java/com/recrutaibackend/repository/EmploymentRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface EmploymentRepository extends JpaRepository<Employment, Long> {
+public interface EmploymentRepository extends JpaRepository<Employment, Integer> {
 }
